requirements.txt
flask
openai
gunicorn
google-cloud-firestore
python-dotenv


